refactor: isolate delay config
