+++ deploy/005-remote-timelock.ts
@@ -2,15 +2,67 @@ import { ethers } from "hardhat";
 import { DeployFunction } from "hardhat-deploy/types";
 import { HardhatRuntimeEnvironment } from "hardhat/types";
 
-import { SUPPORTED_NETWORKS } from "../helpers/deploy/constants";
-import { DelayConfig, delayConfig } from "../helpers/deploy/deploymentConfig";
+export enum REMOTE_NETWORKS {
+  ETHERUEM = "ethereum",
+  OPBNBMAINNET = "opbnbmainnet",
+  ARBITRUM_ONE = "arbitrumone",
+  SEPOLIA = "sepolia",
+  OPBNBTESTNET = "opbnbtestnet",
+  ARBITRUM_SEPOLIA = "arbitrumsepolia",
+  HARDHAT = "hardhat",
+}
+type DelayTypes = {
+  normal: number;
+  fast: number;
+  critical: number;
+};
+export type DelayConfig = {
+  [key in REMOTE_NETWORKS]: DelayTypes;
+};
 
+export const delayConfig: DelayConfig = {
+  hardhat: {
+    normal: 600,
+    fast: 300,
+    critical: 100,
+  },
+  sepolia: {
+    normal: 600,
+    fast: 300,
+    critical: 100,
+  },
+  ethereum: {
+    normal: 172800,
+    fast: 21600,
+    critical: 3600,
+  },
+  opbnbtestnet: {
+    normal: 600,
+    fast: 300,
+    critical: 100,
+  },
+  opbnbmainnet: {
+    normal: 172800,
+    fast: 21600,
+    critical: 3600,
+  },
+  arbitrumsepolia: {
+    normal: 600,
+    fast: 300,
+    critical: 100,
+  },
+  arbitrumone: {
+    normal: 172800,
+    fast: 21600,
+    critical: 3600,
+  },
+};
 const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   const { deployments, getNamedAccounts } = hre;
   const { deploy } = deployments;
   const { deployer } = await getNamedAccounts();
 
-  const networkName = hre.network.name as SUPPORTED_NETWORKS;
+  const networkName = hre.network.name as REMOTE_NETWORKS;
   const live = hre.network.live;
 
   const omnichainGovernanceExecutorAddress = (await ethers.getContract("OmnichainGovernanceExecutor")).address;
@@ -18,7 +70,7 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   await deploy("NormalTimelock", {
     contract: live ? "TimelockV8" : "TestTimelockV8",
     from: deployer,
-    args: [omnichainGovernanceExecutorAddress, (delayConfig as DelayConfig)[networkName].normal],
+    args: [omnichainGovernanceExecutorAddress, delayConfig[networkName].normal],
     log: true,
     autoMine: true,
   });
@@ -26,7 +78,7 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   await deploy("FastTrackTimelock", {
     contract: live ? "TimelockV8" : "TestTimelockV8",
     from: deployer,
-    args: [omnichainGovernanceExecutorAddress, (delayConfig as DelayConfig)[networkName].fast],
+    args: [omnichainGovernanceExecutorAddress, delayConfig[networkName].fast],
     log: true,
     autoMine: true,
   });
@@ -34,7 +86,7 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   await deploy("CriticalTimelock", {
     contract: live ? "TimelockV8" : "TestTimelockV8",
     from: deployer,
-    args: [omnichainGovernanceExecutorAddress, (delayConfig as DelayConfig)[networkName].critical],
+    args: [omnichainGovernanceExecutorAddress, delayConfig[networkName].critical],
     log: true,
     autoMine: true,
   });
